created new test class OutputFile and added/changed some error messages especially in Bake.rb

+++ lib/Converter.rb
@@ -14,17 +14,17 @@ module BConv
     end
     
     def convert
-      begin
-        outputfilename = Util.makeAbsolute(@map['OutputFileName'], @configFile)
-        templatefilename = Util.makeAbsolute(@map['TemplateFile'], @configFile)
-        rescue Exception => msg
-          abort msg
-      end
+      outputfilename = Util.makeAbsolute(@map['OutputFileName'], @configFile)
+      templatefilename = Util.makeAbsolute(@map['TemplateFile'], @configFile)
 
+      lineIdx = 0
+      tmpLineIdx = 0
+      set = true
       begin
       File.open(outputfilename, 'w') do |fout|
         File.open(templatefilename) do |fread|
           File.readlines(fread).each do |line|
+            lineIdx += 1
             wroteLine = false
             @map.keys.each do |key|               
 
@@ -36,8 +36,12 @@ module BConv
                 end
                 
                 filename = Util.makeAbsolute(@map[key], @configFile)
+                raise "Error: Template file #{File.basename(filename)} is empty!" if File.zero?(filename)
+                raise "Error: File #{File.basename(filename)} does not exist!" if !File.exist?(filename)
+                
                 File.open(filename) do |fr|
                   File.readlines(fr).each do |l|
+                    tmpLineIdx += 1
                     @map.keys.each do |k|
                       wroteLine = findAndReplace(k, l, fout, prefix, postfix)
                       break if wroteLine == true
@@ -45,7 +49,12 @@ module BConv
                     if wroteLine == false 
                       l.strip!
                       l = prefix + l + postfix + "\n"
+                      m = l.match(/\$\$\((.*)\)/)
+                      if m
+                        puts "Info: Key $$(#{m[1]}) in #{File.basename(filename)}, line #{tmpLineIdx.to_s} wasn\'t replaced!" if m  
+                      end
                       fout.write(l)
+                      set = false
                     end
                   end
                 end
@@ -53,19 +62,27 @@ module BConv
                 @map.store(key,Util.strToArray(key, @map))
                 wroteLine = findAndReplace(key, line, fout, "", "")
               elsif line.include?(key.to_s)
-                wroteLine = findAndReplace(key, line, fout, "", "")
+                wroteLine = findAndReplace(key, line, fout, "", "")  
               end
             end
-          
-           if wroteLine == false 
-            fout.write(line)
-           end
-                                   
+            
+            if wroteLine == false
+              if set == true
+                m = line.match(/\$\$\((.*)\)/)
+                if m
+                  puts "Info: Key $$(#{m[1]}) in #{File.basename(templatefilename)}, line #{lineIdx.to_s} wasn\'t replaced!"
+                end 
+                fout.write(line)
+              end
+              set = true
+            end 
           end
         end
       end
-      rescue Exception => msg
-        abort msg
+      raise "Error: Output file #{File.basename(outputfilename)} is empty!" if File.zero?(outputfilename)
+      raise "Error: Template file #{File.basename(templatefilename)} is empty!" if File.zero?(templatefilename)
+      rescue Exception => e
+        abort e.message
       end
     end
